Finish function to get url with game ID
url.js
